requirements.txt
flask==2.3.3
firebase-admin==6.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytz==2023.3